+++ FL_MDVRP/FL_MDVRP/DistanceTable.h
@@ -6,28 +6,28 @@ class Graph;
 
 class DistanceTable
 {
-	short int  V;    // No. of vertices
+	int  V;    // No. of vertices
 	float ** a;
 	// In a weighted graph, we need to store vertex
 	// and weight pair for every edge
 	
 
 public:
-	short int const NO_ADJ = -1;
+	int const NO_ADJ = -1;
 
-	DistanceTable(short int V);  // Constructor
+	DistanceTable(int V);  // Constructor
 
 	~DistanceTable(); 
 								 
-	void addEdge(short int u, short int v, short int w); // function to add an edge to graph
+	void addEdge(int u, int v, float w); // function to add an edge to graph
 
-	void addDiagEdge(short int u, short int w);
+	void addDiagEdge(int u, float w);
 
-	float getEdge(short int u, short int v);
+	float getEdge(int u, int v);
 	
 	void loadMinDistanceTable(DistVect * dv); // load the table with the minimum distance among all vertexes
 
-	void loadMinDistanceEdge(DistVect * dv, short int i, short int j); // load minimum distance between i and j
+	void loadMinDistanceEdge(DistVect * dv, int i, int j); // load minimum distance between i and j
 
 	void fillDistanceTable(DistVect * dv, Graph * g); // fill the distance table with the minimum distance between each pair of vertexes
 
@@ -35,11 +35,11 @@ public:
 
 	void fillDistanceTableFromPrevVertex(DistVect * dv, Graph * g);
 
-	void fillFromPrevVertex(DistVect * dv, Graph * g, short int v);
+	void fillFromPrevVertex(DistVect * dv, Graph * g, int v);
 
 	void fillFromPath(DistVect * dv, Graph * g);
 
-	short int getNextSourceForDijkstra();
+	int getNextSourceForDijkstra();
 
 	bool allEdgesFilled();
 };
